created PrintGraphCommand +semver:major
