+++ src/DependencyGraph.App.IntegrationTests/TestFiles/TestSolution.sln_Exclude.dgml
@@ -31,8 +31,8 @@
     <Node Id="WebApplication" Category="ProjectDependency" Label="WebApplication" />
     <Node Id="WebApplication/net8.0" Category="FrameworkDependency" Label="WebApplication/net8.0" />
     <Node Id="AutoMapper@13.0.1" Category="PackageDependency" Label="AutoMapper@13.0.1" />
-    <Node Id="Microsoft.DotNet.ILCompiler@8.0.11" Category="PackageDependency" Label="Microsoft.DotNet.ILCompiler@8.0.11" />
-    <Node Id="Microsoft.NET.ILLink.Tasks@8.0.11" Category="PackageDependency" Label="Microsoft.NET.ILLink.Tasks@8.0.11" />
+    <Node Id="Microsoft.DotNet.ILCompiler@8.0.12" Category="PackageDependency" Label="Microsoft.DotNet.ILCompiler@8.0.12" />
+    <Node Id="Microsoft.NET.ILLink.Tasks@8.0.12" Category="PackageDependency" Label="Microsoft.NET.ILLink.Tasks@8.0.12" />
     <Node Id="WpfApp" Category="ProjectDependency" Label="WpfApp" />
     <Node Id="WpfApp/net8.0-windows7.0" Category="FrameworkDependency" Label="WpfApp/net8.0-windows7.0" />
   </Nodes>
@@ -92,8 +92,8 @@
     <Link Source="WebApplication/net8.0" Target="AutoMapper@13.0.1" />
     <Link Source="WebApplication/net8.0" Target="ClassLibrary" />
     <Link Source="ClassLibrary" Target="AutoMapper@13.0.1" />
-    <Link Source="WebApplication/net8.0" Target="Microsoft.DotNet.ILCompiler@8.0.11" />
-    <Link Source="WebApplication/net8.0" Target="Microsoft.NET.ILLink.Tasks@8.0.11" />
+    <Link Source="WebApplication/net8.0" Target="Microsoft.DotNet.ILCompiler@8.0.12" />
+    <Link Source="WebApplication/net8.0" Target="Microsoft.NET.ILLink.Tasks@8.0.12" />
     <Link Source="WpfApp" Target="WpfApp/net8.0-windows7.0" />
     <Link Source="WpfApp/net8.0-windows7.0" Target="ClassLibrary" />
   </Links>
