+++ src/DependencyGraph.App.IntegrationTests/TestFiles/TestSolution/global.json
@@ -1,0 +1,6 @@
+{
+  "sdk": {
+    "version": "8.0.405",
+    "rollForward": "disable"
+  }
+}
