+++ src/DependencyGraph.App/Properties/launchSettings.json
@@ -2,7 +2,7 @@
   "profiles": {
     "DependencyGraph.App": {
       "commandName": "Project",
-      "commandLineArgs": "..\\..\\..\\DependencyGraph.sln -v dgml -o ..\\..\\..\\DependencyGraph.App\\DependencyGraph.sln.dgml --no-restore -e *Tests*"
+      "commandLineArgs": "trace -h"
     }
   }
 }
